Fjernet logging. Flyttet laget ny fil for 'BrevRequest', fjernet duplikat kode som ble brukt for 'Flettefelter'.
