--- src/main/kotlin/no/nav/familie/ef/sak/brev/dto/SanityBrevRequestInnhentingAktivitetsplikt.kt
@@ -1,10 +1,0 @@
-package no.nav.familie.ef.sak.brev.dto
-
-data class SanityBrevRequestInnhentingAktivitetsplikt(
-    val flettefelter: Flettefelter,
-)
-
-data class Flettefelter(
-    val navn: List<String>,
-    val fodselsnummer: List<String>,
-)
